refactor(files): DIP relocate repo interfaces domain/repositories → application/usecases
Mirror v0.157.1 curriculum / v0.160.1 users / v0.162.1 messaging
DIP-style relocation. The repository ports describe what the use case
LAYER needs from infrastructure, so they belong in `application/usecases`
(consumer), not `domain/repositories` (provider-adjacent).

- New consolidated file: `application/usecases/repository_interfaces.go`
  hosts both `FileMetadataRepository` + `FileVersionRepository` interfaces
  (copied verbatim from the deleted files for stable doc).
- Old files removed: `domain/repositories/{file_metadata,file_version}_repository.go`
  + empty directory.
- 5 import sites updated: file_usecase.go / version_usecase.go drop the
  `repositories` import and use bare type names (same-package now);
  persistence/{file_metadata,file_version}_repository.go switch to
  importing `application/usecases` to declare the constructor's return
  type; file_handler_usecase_test.go compile-time assertion `var _ ...`
  re-qualified.

Domain stays free of port abstractions — only `entities` is imported by
the new file. Pure refactor — no behavior change. All tests pass.

Refs #317
