+++ internal/modules/files/application/usecases/repository_interfaces.go
@@ -1,5 +1,10 @@
-// Package repositories определяет интерфейсы репозиториев модуля files.
-package repositories
+// Package usecases — repository ports for the files module.
+//
+// Repository interfaces live в the consuming package (usecases) per
+// Dependency Inversion Principle. Domain layer stays free of infrastructure
+// abstractions; entities are the only thing imported here. Mirror v0.157.1
+// curriculum + v0.160.1 users + v0.162.1 messaging.
+package usecases
 
 import (
 	"context"
@@ -57,4 +62,34 @@ type FileMetadataRepository interface {
 
 	// CleanupExpired удаляет временные файлы с истёкшим сроком.
 	CleanupExpired(ctx context.Context) (int64, error)
+}
+
+// FileVersionRepository определяет интерфейс доступа к версиям файлов.
+type FileVersionRepository interface {
+	// Create сохраняет новую запись версии файла.
+	Create(ctx context.Context, version *entities.FileVersion) error
+
+	// GetByID получает версию файла по ID.
+	GetByID(ctx context.Context, id int64) (*entities.FileVersion, error)
+
+	// GetByFileMetadataID получает все версии файла.
+	GetByFileMetadataID(ctx context.Context, fileMetadataID int64) ([]*entities.FileVersion, error)
+
+	// GetLatestVersion получает последнюю версию файла.
+	GetLatestVersion(ctx context.Context, fileMetadataID int64) (*entities.FileVersion, error)
+
+	// GetByVersionNumber получает конкретную версию по номеру.
+	GetByVersionNumber(ctx context.Context, fileMetadataID int64, versionNumber int) (*entities.FileVersion, error)
+
+	// Delete удаляет запись версии файла.
+	Delete(ctx context.Context, id int64) error
+
+	// DeleteByFileMetadataID удаляет все версии файла.
+	DeleteByFileMetadataID(ctx context.Context, fileMetadataID int64) error
+
+	// CountByFileMetadataID возвращает количество версий файла.
+	CountByFileMetadataID(ctx context.Context, fileMetadataID int64) (int64, error)
+
+	// GetNextVersionNumber возвращает следующий номер версии для файла.
+	GetNextVersionNumber(ctx context.Context, fileMetadataID int64) (int, error)
 }
